requirements.txt
streamlit>=1.32.0
openpyxl>=3.1.0
pandas>=2.0.0
python-barcode[images]>=0.15.1
Pillow>=10.0.0
